2 - Intermediate/ex_4.py
# Créez un Cookie Clicker en CustomTkinter qui:
# 1. Affiche un gros cookie cliquable au centre
# 2. Compte les cookies en temps réel
# 3. Propose des améliorations à acheter:
#    - Auto-clicker: +1 cookie/sec
#    - Double-click: x2 cookies par clic
#    - Cookie d'or: apparaît aléatoirement, bonus temporaire
# 4. Sauvegarde la progression (cookies, améliorations)
# 5. Affiche des animations lors des clics

# Librairies: customtkinter, PIL pour les images
# Concepts: GUI, gestion d'événements, sauvegarde

import customtkinter as ctk
from PIL import Image
import json
import random
import time

class CookieClicker:
   def __init__(self):
       # Code à compléter...