2 - Intermediate/requirements.txt
customtkinter==5.2.2
Pillow==10.2.0
psutil==5.9.8
sqlite3-utils==3.35.2
cryptography==41.0.7
flask==3.0.2
python-dateutil==2.8.2
click==8.1.7
colorama==0.4.6